Add documentation and reorganize code
