+++ lib/delta-processing.js
@@ -1,5 +1,5 @@
 import { getTypes, getMatchingUuids, getValuesFromPath } from "./queries";
-const CONFIG = require("/config/config.json");
+import { CONFIG } from "../env-config"
 
 /**
  * Gets the query params from the received deltas.
