+++ lib/queries.js
@@ -1,9 +1,12 @@
 import { sparqlEscapeUri } from "mu";
 import { querySudo as query } from "@lblod/mu-auth-sudo";
-import { CONFIG } from "../env-config";
+import {
+  constructGraphsFilter,
+  constructPathInAndOutOfSourceGraphs,
+} from "./utils";
 
 export async function getTypes(subject) {
-  const graphsFilter = constructGraphsFilter();
+  const graphsFilter = constructGraphsFilter("graph");
 
   const queryGetType = `
     SELECT DISTINCT ?type WHERE {
@@ -57,12 +60,14 @@ export async function getValuesFromPath(subject, path, filter) {
     queryFilter = `BIND(${sparqlEscapeUri(subject)} as ?s) \n ${filter}`;
   }
 
-  const pathInAndOutOfSourceGraphs = constructPathInAndOutOfSourceGraphs(path);
-
-  const graphsFilter = constructGraphsFilter();
+  const pathInAndOutOfSourceGraphs = constructPathInAndOutOfSourceGraphs(
+    path,
+    "graph"
+  );
+  const graphsFilter = constructGraphsFilter("graph");
 
   const queryGetValue = `
-    SELECT DISTINCT ?value WHERE {
+    SELECT DISTINCT ?o WHERE {
       ${pathInAndOutOfSourceGraphs}
       ${queryFilter}
       ${graphsFilter}
@@ -72,7 +77,7 @@ export async function getValuesFromPath(subject, path, filter) {
   const result = await query(queryGetValue);
 
   if (result.results.bindings.length) {
-    return result.results.bindings.map((res) => res.value.value);
+    return result.results.bindings.map((res) => res.o.value);
   } else {
     console.log(
       `Resource ${subject} has no value for path ${[
@@ -82,77 +87,4 @@ export async function getValuesFromPath(subject, path, filter) {
     );
     return null;
   }
-}
-
-function constructGraphsFilter() {
-  const escapedSourceGraphs = CONFIG.sourceGraphs.map((sourceGraph) =>
-    sparqlEscapeUri(sourceGraph)
-  );
-  const graphsFilterStr = `FILTER(?graph IN ( ${escapedSourceGraphs.join(
-    ", "
-  )}))`;
-  return graphsFilterStr;
-}
-
-function constructPredicatePath(arrayPath) {
-  let stringifiedPath = arrayPath.map((predicate) => {
-    if (predicate[0] == "^") {
-      return `^${sparqlEscapeUri(predicate.slice(1))}`;
-    } else {
-      return `${sparqlEscapeUri(predicate)}`;
-    }
-  });
-  stringifiedPath = stringifiedPath.join("/");
-  return stringifiedPath;
-}
-
-/**
- * We need to get only the values that are in the predefined source graphs.
- * But sometime the paths spead accross multiple graphs, so we split the path
- * to be able to force the ?value to be in our source graphs.
- *
- * Examples:
- *   1. Only one value in the path ["http://www.w3.org/2004/02/skos/core#prefLabel"]
- *      Output will be:
- *        ```
- *        GRAPH ?graph {
- *          ?subject <http://www.w3.org/2004/02/skos/core#prefLabel> ?value .
- *        }
- *        ```
- *   2. More than one value in the path
- *      ["http://www.w3.org/ns/org#classification", "http://www.w3.org/2004/02/skos/core#prefLabel"]
- *      We only set the latest part of the path in the graph, the rest could be spread over multiple graphs
- *      Output will be:
- *        ```
- *        ?subject <http://www.w3.org/ns/org#classification> ?resource .
- *        GRAPH ?graph {
- *          ?resource <http://www.w3.org/2004/02/skos/core#prefLabel> ?value .
- *        }
- *        ```
- */
-function constructPathInAndOutOfSourceGraphs(path) {
-  // Deep copy of the path to avoid modifying the config object
-  const clonedPath = JSON.parse(JSON.stringify(path))
-  const pathPortionInSourceGraphs = constructPredicatePath([clonedPath.pop()]);
-  const pathPotentiallyOutsideSourceGraphs = constructPredicatePath(clonedPath);
-
-  let pathInAndOutOfSourceGraphs;
-  if (pathPotentiallyOutsideSourceGraphs.length) {
-    pathInAndOutOfSourceGraphs = `
-      ${sparqlEscapeUri(
-        subject
-      )} ${pathPotentiallyOutsideSourceGraphs} ?resource .
-
-      GRAPH ?graph {
-        ?resource ${pathPortionInSourceGraphs} ?value .
-      }
-    `;
-  } else {
-    pathInAndOutOfSourceGraphs = `
-      GRAPH ?graph {
-        ${sparqlEscapeUri(subject)} ${pathPortionInSourceGraphs} ?value .
-      }
-    `;
-  }
-  return pathInAndOutOfSourceGraphs;
 }
